separate files, add email filed and project setting page
